feat: added purge on modularity module save
